+++ AGENTS.md
@@ -23,9 +23,9 @@ open-code-review (`ocr`) is an AI-powered code review CLI tool written in Go (mo
 ## Code Style
 
 - After writing code, run `make check` to format and check the code.
-- `make check` runs: license check, CJK check, `go mod tidy`, `gofmt -s -w .`, and `go vet`.
-- Source files are English-only — comments, identifiers and strings alike. Translated prose belongs in `README.<locale>.md`, `pages/src/content/docs/<locale>/` or an i18n table. `make cjk-check` enforces this in CI; it also rejects fullwidth punctuation (`：`, `（`), which is easy to leave behind in an otherwise English sentence.
-- When CJK text is intentional (an encoding fixture, a language-switcher label), append an `allow-cjk: <reason>` marker comment to that line rather than widening the allowlist in `scripts/verify-cjk.go`.
+- `make check` runs: license check, English-only check, `go mod tidy`, `gofmt -s -w .`, and `go vet`.
+- Source files must not contain unapproved non-English text in comments, identifiers or strings. Translated prose belongs in `README.<locale>.md`, `pages/src/content/docs/<locale>/` or an i18n table. `make english-check` enforces this in CI. It flags every letter outside ASCII, whichever the writing system — Han, kana, Hangul, Cyrillic, Greek, Arabic, and equally the diacritics that spell German, French or Vietnamese — plus combining accents and fullwidth punctuation (`：`, `（`), which is easy to leave behind in an otherwise English sentence. Symbols and emoji (`─ → ≥ ✅`) pass, since they are not letters. One gap remains: another language spelled entirely in ASCII (`Loeschen der Datei`, or a romanised transcription) takes a dictionary to spot, so it stays a matter for review.
+- When non-English text is intentional (an encoding fixture, a language-switcher label), append an `allow-non-english: <reason>` marker comment to that line rather than widening the allowlist in `scripts/verify-english-only.go`.
 
 ## Testing
