+++ examples/gerrit_ci/post_review_test.py
@@ -134,10 +134,10 @@ class BuildReviewInputTest(unittest.TestCase):
         self.assertNotIn("**Suggestion:**", msg)
 
     def test_unicode_comment_roundtrip(self):
-        content = "空指针解引用：y 可能为 nil"  # allow-cjk: fixture exercises UTF-8 comment bodies
-        ri = build([comment(path="pkg/服务.go", content=content)])  # allow-cjk: fixture exercises UTF-8 file paths
-        self.assertIn("pkg/服务.go", ri["comments"])  # allow-cjk: fixture exercises UTF-8 file paths
-        self.assertIn(content, entry_of(ri, "pkg/服务.go")["message"])  # allow-cjk: fixture exercises UTF-8 file paths
+        content = "空指针解引用：y 可能为 nil"  # allow-non-english: fixture exercises UTF-8 comment bodies
+        ri = build([comment(path="pkg/服务.go", content=content)])  # allow-non-english: fixture exercises UTF-8 file paths
+        self.assertIn("pkg/服务.go", ri["comments"])  # allow-non-english: fixture exercises UTF-8 file paths
+        self.assertIn(content, entry_of(ri, "pkg/服务.go")["message"])  # allow-non-english: fixture exercises UTF-8 file paths
         self.assertEqual(json.loads(json.dumps(ri, ensure_ascii=False)), ri)
 
     def test_path_with_spaces(self):
@@ -539,7 +539,7 @@ class MakePosterTest(unittest.TestCase):
     def test_preemptive_basic_auth_and_utf8_body(self):
         import base64
 
-        req, _parsed = self.post({"message": "空指针解引用：y 可能为 nil"})  # allow-cjk: fixture exercises UTF-8 request payloads
+        req, _parsed = self.post({"message": "空指针解引用：y 可能为 nil"})  # allow-non-english: fixture exercises UTF-8 request payloads
         auth = req.get_header("Authorization")
         self.assertIsNotNone(auth, "Authorization header must be set preemptively")
         self.assertTrue(auth.startswith("Basic "))
@@ -547,7 +547,7 @@ class MakePosterTest(unittest.TestCase):
             base64.b64decode(auth[len("Basic "):]).decode("utf-8"),
             "review-bot:s3cret-pass",
         )
-        self.assertIn("空指针解引用".encode("utf-8"), req.data)  # allow-cjk: fixture exercises UTF-8 request payloads
+        self.assertIn("空指针解引用".encode("utf-8"), req.data)  # allow-non-english: fixture exercises UTF-8 request payloads
         self.assertIn("application/json", req.get_header("Content-type"))
 
     def test_xssi_response_parses(self):
