+++ pages/src/components/Footer.tsx
@@ -10,9 +10,9 @@ import type { Language } from '../i18n/types';
 
 const LANG_OPTIONS: { value: Language; label: string }[] = [
   { value: 'en', label: 'English' },
-  { value: 'zh', label: '中文' }, // allow-cjk: language options are labelled in their own language
-  { value: 'ja', label: '日本語' }, // allow-cjk: language options are labelled in their own language
-  { value: 'ru', label: 'Русский' },
+  { value: 'zh', label: '中文' }, // allow-non-english: language options are labelled in their own language
+  { value: 'ja', label: '日本語' }, // allow-non-english: language options are labelled in their own language
+  { value: 'ru', label: 'Русский' }, // allow-non-english: language options are labelled in their own language
 ];
 
 const Footer: React.FC = () => {
